sidebar ve tabHome components klasörüne taşındı.
