--- src/pages/home/homeTypes.ts
@@ -1,2 +1,0 @@
-
-
