--- src/pages/home/tabHome/index.tsx
@@ -1,11 +1,0 @@
-import EventList from "../../../components/eventList"
-
-const TabHome = () => {
-  return (
-    <div>
-        <EventList/>
-    </div>
-  )
-}
-
-export default TabHome
